metabolights-referencelayer/importer/src/main/java/uk/ac/ebi/metabolights/referencelayer/importer/Percentage.java
/*
 * EBI MetaboLights - http://www.ebi.ac.uk/metabolights
 * Cheminformatics and Metabolism group
 *
 * European Bioinformatics Institute (EMBL-EBI), European Molecular Biology Laboratory, Wellcome Trust Genome Campus, Hinxton, Cambridge CB10 1SD, United Kingdom
 *
 * Last modified: 9/24/14 5:05 PM
 * Modified by:   conesa
 *
 *
 * ©, EMBL, European Bioinformatics Institute, 2014.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *    http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */

package uk.ac.ebi.metabolights.referencelayer.importer;

/**
 * User: conesa
 * Date: 24/09/2014
 * Time: 17:05
 */
public class Percentage {

	private int total = 100;
	private int count = 0;
	public Percentage(int total){
		this.total = total;
	}

	public Percentage() {

	}

	public int getTotal() {
		return total;
	}

	public void setTotal(int total) {
		this.total = total;
	}

	public int getCount() {
		return count;
	}

	public void setCount(int count) {
		this.count = count;
	}

	public void oneMore(){
		count = count+1;
	}
	public String getPercentage(){
		return String.valueOf((int)(((double)count/(double)total)*100));
	}
	@Override
	public String toString(){
		return getPercentage() + "%";
	}


	
}
